// webapp/src/types/api/EntityDetailResponse.ts
// This file is auto-generated by `cargo run --bin export_types`
// Do not edit manually.

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Entity } from "../database/Entity";

export type EntityDetailResponse = { entity: Entity, references_count: number, items_count: number, relations_count: number, };